requirements.txt
openai>=1.0.0
python-dotenv>=0.19.0
regex>=2023.0.0
json5>=0.9.14
flask>=3.0.0